requirements_jetson.txt
# Jetson Orin Nano Requirements
# =============================
# Agricultural RAG System için optimize edilmiş paketler

# Core ML Framework
torch==2.0.0
torchvision==0.15.0
torch-tensorrt==1.4.0

# Transformers Ecosystem
transformers==4.35.0
tokenizers==0.14.0
datasets==2.14.0
accelerate==0.24.0
optimum[onnxruntime]==1.13.0

# RAG Components
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# ONNX Runtime
onnxruntime-gpu==1.16.0

# Scientific Computing
numpy==1.24.3
scipy==1.11.0
pandas==2.1.0
scikit-learn==1.3.0

# System Monitoring
psutil==5.9.5
gpustat==1.1.1
nvidia-ml-py==12.535.108

# Visualization
matplotlib==3.7.0
seaborn==0.12.0
tqdm==4.66.0

# Development
jupyter==1.0.0
ipykernel==6.25.0
black==23.7.0
flake8==6.0.0
pytest==7.4.0

# Additional utilities
Pillow==10.0.0
requests==2.31.0
urllib3==2.0.4 